mend

+++ models.py
@@ -1,11 +1,17 @@
+import os
+import urllib.parse
 from peewee import (SqliteDatabase, IntegerField, DoubleField,
                     DateTimeField, datetime as peewee_datetime, Model,
-                    CharField, TextField)
+                    CharField, TextField, PostgresqlDatabase)
 
 from config import DB_NAME
 
-
-db = SqliteDatabase(DB_NAME)
+if os.environ.get("DATABASE_URL"):
+    url = urllib.parse.urlparse(os.environ.get("DATABASE_URL"))
+    db = PostgresqlDatabase(host=url.hostname, user=url.username, password=url.password, post=url.port,
+                            database=url.path[1:])
+else:
+    db = SqliteDatabase(DB_NAME)
 
 
 class _Model(Model):
@@ -60,17 +66,18 @@ class ErrorLog(_Model):
     created = DateTimeField(default=peewee_datetime.datetime.now, index=True)
 
 
-def init_db():
-    Rate.drop_table()
-    Rate.create_table()
-    Rate.create(from_currency=840, to_currency=980, rate=1, module="privat_api")
-    Rate.create(from_currency=840, to_currency=643, rate=1, module="cbr_api")
-    Rate.create(from_currency=1000, to_currency=840, rate=1, module="privat_api")
-    Rate.create(from_currency=1000, to_currency=980, rate=1, module="cryptonator_api")
-    Rate.create(from_currency=1000, to_currency=643, rate=1, module="cryptonator_api")
+def start_db():
+    if not Rate.table_exists():
+        Rate.drop_table()
+        Rate.create_table()
+        Rate.create(from_currency=840, to_currency=980, rate=1, module="privat_api")
+        Rate.create(from_currency=840, to_currency=643, rate=1, module="cbr_api")
+        Rate.create(from_currency=1000, to_currency=840, rate=1, module="privat_api")
+        Rate.create(from_currency=1000, to_currency=980, rate=1, module="cryptonator_api")
+        Rate.create(from_currency=1000, to_currency=643, rate=1, module="cryptonator_api")
 
-    for m in (ApiLog, ErrorLog):
-        m.drop_table()
-        m.create_table()
+        for m in (ApiLog, ErrorLog):
+            m.drop_table()
+            m.create_table()
 
-    print("db created!")
+        print("db created!")
